fix(desktop-shell-ball): move dashboard route navigation out of platform
